KB regenerate
git-svn-id: svn://svn.liferay.com/repos/public/plugins/trunk@52677 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portlets/knowledge-base-portlet/docroot/WEB-INF/service/com/liferay/knowledgebase/model/ArticleClp.java
@@ -136,11 +136,11 @@ public class ArticleClp extends BaseModelImpl<Article> implements Article {
 		_parentResourcePrimKey = parentResourcePrimKey;
 	}
 
-	public double getVersion() {
+	public int getVersion() {
 		return _version;
 	}
 
-	public void setVersion(double version) {
+	public void setVersion(int version) {
 		_version = version;
 	}
 
@@ -380,7 +380,7 @@ public class ArticleClp extends BaseModelImpl<Article> implements Article {
 	private Date _createDate;
 	private Date _modifiedDate;
 	private long _parentResourcePrimKey;
-	private double _version;
+	private int _version;
 	private String _title;
 	private String _content;
 	private String _description;
